impl Feedback Retrieval Service and build performance containers and its dependencies

+++ src/main/java/com/cob/feedback/model/performance/ClinicalContainer.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class ClinicalContainer {
     private long happyIndex;
     private long nps;
-    private long average;
+    private double average;
 }
